Split automation parity gates
